Modificación tipo de dato del id de TransactionDetail

+++ src/main/java/com/inventario/Inventario/entities/TransactionDetail.java
@@ -14,7 +14,7 @@ public class TransactionDetail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Setter(AccessLevel.NONE)
-    private Long id; // Cambiar a Integer
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "product_id", nullable = false)
